requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
python-dotenv==1.0.1